pyproject.toml
[tool.poetry]
name = "vid-helper"
version = "0.1.0"
description = ""
authors = ["Zeeshan Khan <zkhan1093@gmail.com>"]
readme = "README.md"
packages = [{include = "vid_helper"}]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.14"
ffmpeg-python = "^0.2.0"
google-api-python-client = "^2.70.0"
google-auth-oauthlib = "^0.8.0"
google-auth-httplib2 = "^0.1.0"
celery = {extras = ["redis"], version = "^5.2.7"}
pydantic = "^1.10.2"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
itsdangerous = "^2.1.2"
cryptography = "^38.0.4"
whisper = {git = "https://github.com/openai/whisper.git"}
torch = "^1.13.1"
tqdm = "^4.64.1"
numpy = "^1.23.5"
transformers = "^4.25.1"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
